requirements.txt
im-data-manager-job-decoder == 1.0.0
munch == 2.5.0
pyyaml == 6.0
